Collect references before starting the object traversal
This provides a better separation of concerns, which will be taken
advantage of shortly.
